Keep setting stack definition using concrete types.
